outsource IterableTaskEncodingDataset, TaskEncoding, TaskEncodingDataset, and TaskEncodingSequence from taskmodule to taskencoding
